cleaned outputs from python processing, and made python variable formatting same
