+++ python_processing/processing_functions/SpaceTimePlotAnalysis.py
@@ -1,0 +1,259 @@
+"""
+    
+SpaceTimePlotAnalysis.py
+
+This script contains the helper functions to analyze space-time plots
+generated from ARFI-SWEI data generated by the sequences in the 
+RSNA-QIBA-US-SWS repository.
+
+Author: Kaden Bock (04/08/2025)
+    
+"""
+
+import numpy as np
+
+# %% Group Velocity Analysis
+
+def run_TTP_to_determine_group_vel(plane:np.array, t_ms:np.array, lat_mm:np.array,
+                                   time_axis_num:int=1):
+    """ Runs the Time-to-Peak (TTP) algorithm to determine the group velocity speed
+        for the input space-time plot.
+    
+    Args:
+        plane (np.array): The space-time plot of the data over which to run the algorithm.
+        t_ms (np.array): The time vector of the space-time plot.
+        lat_mm (np.array): The lateral position vector of the space-time plot.
+        time_axis_num (int, optional): The axis number corresponding to the time axis of
+            The space-time plot. Defaults to 1. The other axis is assumed to be the
+            lateral position axis.
+
+    Returns:
+        np.array: A vector containing the time positions of all peaks.
+        np.float: The group shear wave speed, in meters per second.
+    """
+    
+    if plane.shape[time_axis_num] != t_ms.shape[0]:
+        raise ValueError("The length of the time axis and the axis over which temporal"+
+                         " information in the plane are not the same. Please"+
+                         " double check and correct your inputs.")
+    
+    TTP_inds = np.argmax(plane, axis=time_axis_num)
+    
+    TTP = t_ms[TTP_inds]
+    
+    p = np.polyfit(lat_mm, TTP, deg=1)
+    
+    gSWS = 1/p[0]
+    
+    return TTP, gSWS
+
+def run_radon_sum_to_determine_group_vel(spctm, lat, t, latmin, latmax, tmin, tmax, N:int=512, 
+                                         speedonly:bool=True, interpkind='cubic', plot:bool|Axes=False):
+    """Calculate the shear wave speed using radon sum algorithm
+    
+    Parameters:
+    ----
+    `spctm`: space time plot with displacement/velocity data [MxN]
+    `lat`: length M lateral location vector of every row of `spctm` in mm
+    `t`: length N time vector in ms
+    `latmin`: minimum value of lateral extent to evaluate
+    `latmax`: maximum value of lateral extent to evaluate
+    `tmin`: minimum value of temporal extent to evaluate
+    `tmax`: maximum value of temporal extent to evaluate
+    `N`: Number of time points to evaluate along lateral and temporal extents
+    `speedonly`: determines whether to return just speed or (speed, radonsum, t)
+
+    Returns:
+    ----
+    `speed`: the estimated speed in m/s
+    or 
+    `(speed, radonsum, t)`: estimated speed in m/s, energy integration chart, and time vector along the edges
+    """
+    from scipy.interpolate import interp1d as i1d
+    import numpy as np
+
+    t = np.array(t).flatten()
+    trange = np.linspace(tmin, tmax, N)
+ 
+    # Find the number of lateral positions contained within given range
+    lat = np.array(lat).flatten()
+    latmask = (lat >= latmin) & (lat <= latmax)
+    lat = lat[latmask]
+    nlat = np.sum(latmask)
+    if nlat == 0:
+        raise ValueError(f"Latmask range of {latmin} to {latmax} does not contain any points")
+    latmin = np.min(lat)
+    latmax = np.max(lat)
+    spctm = spctm[latmask,:]
+
+    t1 = np.tile(trange.reshape((-1,1)), (1, N))
+    t2 = np.tile(trange.reshape((1,-1)), (N, 1))
+
+    sums = np.zeros((N,N))
+    for ilat in range(nlat):
+        tsel = (lat[ilat] - latmin)*(t2-t1)/(latmax-latmin) + t1
+        f = i1d(t, spctm[ilat,:], kind=interpkind, bounds_error=False, fill_value=0)
+        sums += f(tsel)
+
+    # limit latsums for ridiculously high speeds
+    dlat = latmax-latmin
+    maxspeed = 15
+    min_dt = dlat/maxspeed
+
+    # start_out, end_out = np.where((t2-t1) < min_dt)
+    # print(start_out,end_out)
+    sums[(t2-t1) <= min_dt] = 0
+    sums[t1 >= t2] = 0
+    st1, st2 = np.where(sums == np.max(sums))
+    st1 = st1[0]
+    st2 = st2[0]
+
+    validst1 = (st1 >= 1) | (st1 < N-1)
+    validst2 = (st2 >= 1) | (st2 < N-1)
+    if not (validst1 and validst2):
+        raise Exception(f"Both temporal coordinates must be between 1 and {N-1}, but found {st1} and {st2}")
+    
+    # estimate the sub-index peak
+    try:
+        Ts = trange[1]-trange[0]
+        peakst1 = sums[(st1-1):(st1+2), st2]
+
+        a = (peakst1[0] + peakst1[2])/2 - peakst1[1]
+        b = (peakst1[2] - peakst1[0])/2
+        dt1 = Ts * -b/(2*a)
+ 
+        peakst2 = sums[st1, (st2-1):(st2+2)]
+        a = (peakst2[0] + peakst2[2])/2 - peakst2[1]
+        b = (peakst2[2] - peakst2[0])/2
+        dt2 = Ts * -b/(2*a)
+
+    except Exception as e: 
+        dt1 = 0
+        dt2 = 0
+    
+    outt2 = trange[st2] + dt2
+    outt1 = trange[st1] + dt1
+    dt12 = np.mean(trange[st2] + dt2 - trange[st1] - dt1)
+    c = float((latmax-latmin)/dt12)
+
+    plot=False
+    if plot:
+        import matplotlib.pyplot as plt
+
+        extent = (t[0], t[-1], lat[-1], lat[0])
+        fig, (ax1,ax2) = plt.subplots(1,2)
+
+        # show the spacetime plot
+        ax1.imshow(spctm,vmin=-.2,vmax=.2,extent=extent,aspect=3)
+
+        # show the processing ranges:
+        ax1.vlines(x=[tmin, np.min([tmax,t[-1]])], ymin=extent[2], ymax=extent[3], color='black')
+        ax1.hlines(y=[latmin, latmax], xmin=extent[0], xmax=extent[1], color='red')
+
+        # Plot the linear fit:
+        # use x = trange
+
+        m = (latmax-latmin)/(outt2-outt1)
+        b = latmin - m*outt1
+
+        y = m*trange + b
+
+        masked_y = np.ma.masked_where((y <= latmin) | (y >= latmax), y)
+        ax1.plot(trange, masked_y, color='blue',label="SWS = {:.2f}".format(c))
+        ax1.legend(loc='best')
+        ax1.set_xlabel("Time (ms)")
+        ax1.set_ylabel("Lateral (mm)")
+
+        ax2.imshow(sums, extent = [tmin,tmax,tmax,tmin])
+        ax2.set_xlabel('Start Time (ms)')
+        ax2.set_ylabel('End Time (ms)')
+        # fig.colorbar(mappable=None,location='right')
+        ax2.scatter(outt2,outt1, color='w')
+        ax2.set_title('Radon sum')
+
+        plt.show() #type:ignore
+
+    if speedonly:
+        return c
+    else:
+        return c, sums, trange, outt1, outt2
+# %% Phase Velocity Analysis
+
+def calculate_phase_vel_from_2D_FT(plane:np.array, t_ms:np.array, lat_mm:np.array,
+                                   freqs_to_analyze:np.array, num_bins:int=2000):
+    """ Calculates the phase velocities for all input frequencies.
+    
+    **NOTE:** This function assumes that lateral positions are indexed in the 0th axis, and
+        temporal positions are indexed in the 1st axis.
+
+    Args:
+        plane (np.array): The space-time plot over which phase information should be calculated.
+        t_ms (np.array): A vector corresponding to the temporal positioning information in the
+            space-time plot.
+        lat_mm (np.array): A vector corresponding to the lateral positioning information in
+            the space-time plot.
+        freqs_to_analyze (np.array): A vector containing all of the frequencies over which the
+            phase velocities will be extracted.
+        num_bins (int, optional): The number of bins which will be used to construct the 2D FFT. 
+            Defaults to 2000.
+
+    Raises:
+        ValueError: In the event that the supplied vector shapes do not match the appropriate axises
+            in the plane, or in the event that a requested frequency to extract phase velocity falls
+            outside of the range of frequencies that are contained within the 2D FFT.
+
+    Returns:
+        np.array: A vector containing all of the extracted phase velocities at all frequencies.
+    """
+    
+    if plane.shape[0] != lat_mm.shape[0]:
+        raise ValueError("The length of the lateral vector and the axis over which lateral"+
+                         " information in the plane is stored are not the same. Please"+
+                         " double check and correct your inputs.")
+    
+    if plane.shape[1] != t_ms.shape[0]:
+        raise ValueError("The length of the time vector and the axis over which temporal"+
+                         " information in the plane is stored are not the same. Please"+
+                         " double check and correct your inputs.")
+        
+    bin_vec = np.arange(-num_bins/2, num_bins/2, 1)
+    
+    dt = np.mean(np.diff(t_ms))/1000 # convert dt from ms to s
+    fs = 1/dt
+    f_t_vec = bin_vec*(fs/num_bins)
+    
+    dlat = np.mean(np.diff(lat_mm))/1000 # convert dlat from mm to m
+    fs_lat = 1/dlat
+    f_lat_vec = bin_vec*(fs_lat/num_bins)
+    
+    fft = np.fft.fft2(plane, s=(len(f_lat_vec), len(f_t_vec)))
+    
+    abs_fft = np.abs(np.fft.fftshift((fft)))
+    
+    phase_vels = np.zeros(len(freqs_to_analyze))
+    
+    for ind, freq in enumerate(freqs_to_analyze):
+        
+        if freq < f_t_vec[0] or freq > f_t_vec[-1]:
+            raise ValueError("The requested frequency to analyze falls outside "+
+                             "of the range of frequencies contained within the 2D fft.")
+        
+        f_t_bin_idx = np.argmin(np.abs(f_t_vec-freq))
+        
+        profile = abs_fft[:, f_t_bin_idx]
+        
+        f_lat_bin_idx = np.argmax(profile)
+        
+        if f_lat_bin_idx == 0 or f_lat_bin_idx==(len(f_lat_vec) -1):
+            max_spacial_freq_val = f_lat_vec[f_lat_bin_idx]
+        else: 
+            y1 = profile[f_lat_bin_idx-1]
+            y2 = profile[f_lat_bin_idx]
+            y3 = profile[f_lat_bin_idx+1]
+            
+            interpolated_val = (y1-y3)/(y1-2*y2+y3)/2
+            max_spacial_freq_val = f_lat_vec[f_lat_bin_idx] + interpolated_val*dlat
+   
+        phase_vels[ind] = -1* freq/max_spacial_freq_val
+
+    return phase_vels
