+++ python_processing/requirements.txt
@@ -1,0 +1,8 @@
+numpy==2.2.4
+pymatreader==1.0.0
+ipykernel==6.29.5
+matplotlib==3.10.1
+scipy==1.15.2
+tqdm==4.67.1
+ipywidgets==8.1.6
+IProgress==0.4
